# Conflicts:
#	src/main/java/com/project/bit/chat/controller/ChatController.java
#	src/main/java/com/project/bit/userStatus/controller/UserStatusController.java
#	src/main/java/com/project/bit/userStatus/mapper/UserStatusMapper.java
#	src/main/java/com/project/bit/userStatus/service/UserStatusService.java
#	src/main/java/com/project/bit/userStatus/service/UserStatusServiceImpl.java
#	src/main/resources/static/css/style.css
#	src/main/resources/static/js/chat/chat.js
#	src/main/resources/templates/fragments/footer.html
#	src/main/resources/templates/fragments/header.html
#	src/test/java/com/project/bit/chat/mapper/MessageMapperTest.java
#	src/test/java/com/project/bit/chat/mapper/ParticipationMapperTest.java
#	src/test/java/com/project/bit/userStatus/mapper/UserStatusMapperTest.java
#	src/test/java/com/project/bit/userStatus/service/UserStatusServiceTest.java

+++ src/main/resources/static/js/project/projectIssueDetail.js
@@ -108,22 +108,28 @@ function issueUpdate() {
 
 function issueFileDelete(){
 	$('#issue-file-delete').on('click', function(e){
-		$.ajax({
-			url: '/issue/delete/file/'+$('#issue-name').data('code'),
-			type: 'DELETE',
-			data: {issuePath: $('#issue-file-item').data('path')}
-		}).done(function(){
-			location.reload();
-		}).fail(function(){
-			alert('issue file delete fail')
-		})
+		if(confirm('삭제하시겠습니까?')){
+			$.ajax({
+				url: '/issue/delete/file/'+$('#issue-name').data('code'),
+				type: 'DELETE',
+				data: {issuePath: $('#issue-file-item').data('path')}
+			}).done(function(){
+				location.reload();
+			}).fail(function(){
+				alert('issue file delete fail')
+			})			
+		}
+		
+		
 		
 	})
 }
 
 function issueDelete(){
 	$('#issue-delete-btn').on('click', function(){
-		location.href='/removeProjectIssue/'+$('#issue-name').data('code');
+		if(confirm('삭제하시겠습니까?')){
+			location.href='/removeProjectIssue/'+$('#issue-name').data('code');			
+		}
 		/*$.ajax({
 			url: '/issue/delete/'+$('#issue-name').data('code'),
 			type: 'DELETE'
